+++ jaxrs/src/test/java/de/taimos/dvalin/jaxrs/annotationscanner/AbstractApi.java
@@ -1,0 +1,9 @@
+package de.taimos.dvalin.jaxrs.annotationscanner;
+
+@Anno1
+public abstract class AbstractApi {
+
+    @Anno2
+    public abstract void foobar();
+
+}