+++ jaxrs/src/test/java/de/taimos/dvalin/jaxrs/annotationscanner/Anno1.java
@@ -1,0 +1,12 @@
+package de.taimos.dvalin.jaxrs.annotationscanner;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.TYPE, ElementType.METHOD})
+public @interface Anno1 {
+    //
+}