+++ jaxrs/src/test/java/de/taimos/dvalin/jaxrs/annotationscanner/ApiImpl.java
@@ -1,0 +1,19 @@
+package de.taimos.dvalin.jaxrs.annotationscanner;
+
+public class ApiImpl extends AbstractApi implements IApi {
+
+    @Override
+    public void foobar() {
+
+    }
+
+    @Override
+    public void barbaz() {
+
+    }
+
+    @Anno1
+    public void blubb() {
+
+    }
+}