+++ jaxrs/src/test/java/de/taimos/dvalin/jaxrs/annotationscanner/IApi.java
@@ -1,0 +1,9 @@
+package de.taimos.dvalin.jaxrs.annotationscanner;
+
+@Anno2
+public interface IApi {
+
+    @Anno1
+    void barbaz();
+
+}